+++ src/main/java/org/fest/assertions/api/android/util/SparseArrayAssert.java
@@ -6,12 +6,12 @@ import org.fest.assertions.api.AbstractAssert;
 import static org.fest.assertions.api.Assertions.assertThat;
 
 /** Assertions for {@link SparseArray} instances. */
-public class SparseArrayAssert extends AbstractAssert<SparseArrayAssert, SparseArray> {
-  public SparseArrayAssert(SparseArray actual) {
+public class SparseArrayAssert<E> extends AbstractAssert<SparseArrayAssert<E>, SparseArray<E>> {
+  public SparseArrayAssert(SparseArray<E> actual) {
     super(actual, SparseArrayAssert.class);
   }
 
-  public SparseArrayAssert hasKey(int key) {
+  public SparseArrayAssert<E> hasKey(int key) {
     isNotNull();
     assertThat(actual.get(key)) //
         .overridingErrorMessage("Expected key <%s> to be present but was not present.", key) //
@@ -19,7 +19,7 @@ public class SparseArrayAssert extends AbstractAssert<SparseArrayAssert, SparseA
     return this;
   }
 
-  public SparseArrayAssert hasSize(int size) {
+  public SparseArrayAssert<E> hasSize(int size) {
     isNotNull();
     int actualSize = actual.size();
     assertThat(actualSize) //