+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Change Log
 ==========
 
+Version 1.X.X *(In Development)*
+--------------------------------
+
+ * New assertions:
+   * `Instrumentation.ActivityMonitor`
+   * `Instrumentation.ActivityResult`
+ * Added convenience methods for `Display`, `LinearLayout`, and `GridLayout`
+   for orientations.
+ * Added error and ellipsis checks for `TextView`.
+ * Support for generics by assertions classes (e.g., `LruCacheAssert<K, V>`.
+
+
 Version 1.0.0 *(2013-01-17)*
 ----------------------------