+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Version 1.X.X *(In Development)*
  * Added convenience methods for orientations to `Display`, `LinearLayout`, and
    `GridLayout`.
  * Added error and ellipsis checks for `TextView`.
- * Support for generics by assertions classes (e.g., `LruCacheAssert<K, V>`.
+ * Support for generics by assertions classes (e.g., `LruCacheAssert<K, V>`).
 
 
 Version 1.0.0 *(2013-01-17)*